+++ SguanFOC库v3.1.0(无感浮点库...writing)/Sguan_DOB.h
@@ -25,11 +25,11 @@ typedef struct{
     SMDO_STRUCT smdo;       // (结构体)超螺旋滑模DOB
     
     uint8_t Pn;             // (电机数据)极对数
-    double Flux;            // (电机数据)磁链
-    double B;               // (电机数据)粘性阻尼
-    double J;               // (电机数据)转动惯量
+    float Flux;             // (电机数据)磁链
+    float B;                // (电机数据)粘性阻尼
+    float J;                // (电机数据)转动惯量
     
-    double T;               // (参数设计)离散周期
+    float T;                // (参数设计)离散周期
     float K1;               // (参数设计)比例项增益
     float K2;               // (参数设计)积分项增益
