+++ SguanFOC库v3.1.0(无感浮点库...writing)/Sguan_Filter.c
@@ -13,17 +13,21 @@
 
 /**
  * @description: 二阶低通滤波器参数初始化
+ * @reminder: (初始化相关系数float->double->float)
+ * @reminder: (单浮点转double运算，提高系数精度)
  * @param {LPF_STRUCT} *lpf
  * @return {*}
  */
 void LPF_Init(LPF_STRUCT *lpf){
-    double temp1 = lpf->T*lpf->Wc*((double)Value_2_SQRT2);
-    double temp2 = lpf->T*lpf->T*lpf->Wc*lpf->Wc;
-    lpf->filter.num[0] = (float)(temp2/(temp2+temp1+4));
-    lpf->filter.num[1] = (float)((2*temp2)/(temp2+temp1+4));
-    lpf->filter.num[2] = (float)(temp2/(temp2+temp1+4));
-    lpf->filter.den[0] = (float)(-8+2*temp2)/(temp2+temp1+4);
-    lpf->filter.den[1] = (float)(temp2-temp1+4)/(temp2+temp1+4);
+    double temp1 = ((double)lpf->T)*((double)lpf->Wc)*((double)Value_2_SQRT2);
+    double temp2 = ((double)lpf->T)*((double)lpf->T)*
+                    ((double)lpf->Wc)*((double)lpf->Wc);
+    lpf->filter.num[0] = (float)(temp2/(temp2+temp1+4.0));
+    lpf->filter.num[1] = (float)((2.0*temp2)/(temp2+temp1+4.0));
+    lpf->filter.num[2] = (float)(temp2/(temp2+temp1+4.0));
+    lpf->filter.den[0] = (float)(-8.0+2.0*temp2)/(temp2+temp1+4.0);
+    lpf->filter.den[1] = (float)(temp2-temp1+4.0)/(temp2+temp1+4.0);
+    
     // 初始化为零
     lpf->filter.i[0] = 0.0f;
     lpf->filter.i[1] = 0.0f;
